// src/Models/CumulativeScores.ts
export interface CumulativeScores {
    tricks: number;
    trumpSuit: string;
    playerOneScore: number;
    playerOneCumulativeScore: number;
    playerTwoScore: number;
    playerTwoCumulativeScore: number;
    playerThreeScore: number;
    playerThreeCumulativeScore: number;
    playerFourScore: number;
    playerFourCumulativeScore: number;
}